pyproject.toml
[project]
name = "job-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Lenny Lopes (lopeslen)", email = "lopeslen.students@zhaw.ch" }
]
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.1",
    "requests>=2.32.4",
]

[project.scripts]
job-tracker = "job_tracker:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
